refactor(storage): rename pagination_handler to pagination
