+++ contracts/cw-storage/src/lib.rs
@@ -3,7 +3,7 @@ pub mod crypto;
 mod cursor;
 mod error;
 pub mod msg;
-mod pagination_handler;
+mod pagination;
 pub mod state;
 
 pub use crate::error::ContractError;
